Added other methods for handling requests

+++ warehouse_app/src/main/java/com/asasan/warehousemanagement/app/repository/ItemRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ItemRepository extends CrudRepository<Item, Long> {
+public interface ItemRepository extends CrudRepository<Item, Integer> {
 
 }
